perf: compact cdef frame filter params

+++ internal/av1/cdef/filter_avx2_amd64_test.go
@@ -34,12 +34,12 @@ func TestFilterBlockAVX2MatchesPureGo(t *testing.T) {
 					for _, sec := range cdefSecondaryStrengthCorpus(coeffShift) {
 						for _, damping := range []int{3 + coeffShift, 5 + coeffShift, 6 + coeffShift} {
 							params := BlockFilterParams{
-								PrimaryStrength:   pri,
-								SecondaryStrength: sec,
-								Direction:         dir,
-								PrimaryDamping:    damping,
-								SecondaryDamping:  damping,
-								CoeffShift:        coeffShift,
+								PrimaryStrength:   uint8(pri),
+								SecondaryStrength: uint8(sec),
+								Direction:         uint8(dir),
+								PrimaryDamping:    uint8(damping),
+								SecondaryDamping:  uint8(damping),
+								CoeffShift:        uint8(coeffShift),
 								Width:             8,
 								Height:            8,
 							}
